Interpret ansi-colors in error buffer
Current versions of elm-format use ANSI codes to display colored output.

Before this change, the raw output of elm-format was being displayed to the
user using the `message` function, which resulted in escape sequences being
shown.

This commit use `ansi-color` to interpret color codes when reading the output
of elm-format, so that when the error message is displayed we only show
meaninful text.